(client/src/components/Navbar): the righest section of the navbar now is a new component called SubNavbarRight. It contains the logic for showing the correct links
