+++ client/src/components/Navbar/SubNavbarRight/SubNavbarRight.module.css
@@ -1,0 +1,77 @@
+.container {
+	display: flex;
+	gap: 15px;
+}
+
+.signContainer {
+	display: flex;
+	align-items: center;
+	gap: 8px;
+	margin-right: 8px;
+	text-decoration: none;
+}
+
+.userGreet {
+	color: #fff;
+	font-size: 14px;
+	margin: 0;
+}
+
+.userPhoto {
+	color: #fff;
+}
+
+.userGreet:hover,
+.userPhoto:hover {
+	color: #c9c9c9;
+}
+
+.notSignContainer {
+	display: flex;
+	align-items: center;
+	gap: 15px;
+}
+
+.createAccountLink,
+.signInLink,
+.cartLink {
+	font-size: 14px;
+	color: #fff;
+	text-decoration: none;
+	transition: 0.2s;
+}
+
+.createAccountLink:hover,
+.signInLink:hover,
+.cartLink:hover {
+	color: #c9c9c9;
+}
+
+.cartLink {
+	font-size: 22px;
+}
+
+.cartSpan {
+	position: absolute;
+	right: -5px;
+	top: 0;
+	width: 15px;
+	height: 15px;
+	border-radius: 50%;
+	font-size: 14px;
+	font-weight: 600;
+	color: white;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	background-color: red;
+}
+
+.logout {
+	color: #fff;
+	cursor: pointer;
+}
+
+.logout:hover {
+	color: #c9c9c9;
+}
